finance/requirements.txt
django
djangorestframework
langchain
openai
nltk
pydantic
